requirements.txt
openpyxl==3.1.5
Pillow==10.4.0
PyQt5==5.15.11
PyQt5_sip==12.15.0
pytesseract==0.3.13
transformers==4.44.1
